fix: Tests are not rerun, GCP test moved, Correcting rounding errors.
