+++ taikun/provider_tests/data_source_taikun_images_gcp_test.go
@@ -1,4 +1,4 @@
-package cc_gcp_tests
+package provider_tests
 
 import (
 	"fmt"
