+++ taikun/provider_tests/resource_taikun_cloud_credential_gcp_test.go
@@ -1,4 +1,4 @@
-package cc_gcp_tests
+package provider_tests
 
 import (
 	"context"
